docs(readme): update readme to beta
A complete beta version of the README is uploaded.
The README includes installation instructions, how to run tests
and how to test the CLI.

The Jupyter notebook is explicitely referenced below Usage example.

It is possible to run tests now, although is a dummy test.
Nevertheless, there are two new tests in the mod/test branch that
will also run with the same command whenever merged.

CONTRIBUTE.MD was removed at the moment, as well as other more
specific sections.

I am happy to receive contributions as we need to merge this
quickly